requirements.txt
amqp==5.3.1
annotated-types==0.7.0
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.8.3
chardet==3.0.4
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
decorator==4.4.2
dnspython==2.8.0
ecdsa==0.19.1
Flask==3.1.2
flask-cors==6.0.1
fpdf==1.7.2
git-filter-repo==2.47.0
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
lxml==6.0.1
MarkupSafe==3.0.2
moviepy==1.0.3
numpy==2.3.2
packaging==25.0
pillow==11.3.0
proglog==0.1.12
prompt_toolkit==3.0.52
psutil==7.0.0
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pymongo==4.14.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
redis==6.4.0
requests==2.32.5
rfc3986==1.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
reportlab
yt-dlp